pyproject.toml
# ============================================================================
# BUILD SYSTEM
# ============================================================================
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

# ============================================================================
# PROJECT METADATA
# ============================================================================
[project]
name = "biorempp-web"
version = "1.0.0"
description = "BioRemPP Web - Bioremediation Potential Profile"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "BioRemPP Development Team"}
]
keywords = ["bioremediation", "bioinformatics", "dash", "plotly", "multi-omics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Dash",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

# ============================================================================
# CORE DEPENDENCIES (PRODUCTION)
# ============================================================================
dependencies = [
    # Web Framework
    "dash[diskcache]>=3.3.0,<4.0.0",
    "dash-bootstrap-components>=2.0.0,<3.0.0",
    "dash-ag-grid>=32.3.0,<33.0.0",
    "Flask>=3.0.3,<4.0.0",
    "Werkzeug>=3.0.6,<4.0.0",
    
    # Data Processing
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.26.0,<2.0.0",
    "openpyxl>=3.1.0,<4.0.0",
    
    # Scientific Computing
    "scipy>=1.14.1,<2.0.0",
    "scikit-learn>=1.5.2,<2.0.0",
    
    # Visualization
    "plotly>=5.18.0",
    "matplotlib>=3.9.3,<4.0.0",
    "UpSetPlot>=0.9.0,<1.0.0",
    
    # Networking and Graph
    "networkx>=3.4.2,<4.0.0",
    
    # Core Utils
    "PyYAML>=6.0.3,<7.0.0",
    "requests>=2.32.3,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "click>=8.1.7,<9.0.0",
    "Jinja2>=3.1.4,<4.0.0",
    "typing-extensions>=4.12.2",
    
    # JSON Processing
    "ujson>=5.8.0,<6.0.0",
    
    # Additional Dependencies
    "ansi2html>=1.9.2,<2.0.0",
    "blinker>=1.9.0,<2.0.0",
    "certifi>=2024.8.30",
    "charset-normalizer>=3.4.0,<4.0.0",
    "contourpy>=1.3.1,<2.0.0",
    "cycler>=0.12.1,<1.0.0",
    "et-xmlfile>=2.0.0,<3.0.0",
    "fonttools>=4.55.2,<5.0.0",
    "idna>=3.10,<4.0.0",
    "importlib-metadata>=8.5.0,<9.0.0",
    "itsdangerous>=2.2.0,<3.0.0",
    "joblib>=1.4.2,<2.0.0",
    "kiwisolver>=1.4.7,<2.0.0",
    "MarkupSafe>=3.0.2,<4.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
    "packaging>=24.2,<25.0.0",
    "pillow>=11.0.0",
    "pyparsing>=3.2.0,<4.0.0",
    "python-dateutil>=2.9.0,<3.0.0",
    "pytz>=2024.2",
    "retrying>=1.3.4,<2.0.0",
    "setuptools>=75.6.0",
    "six>=1.17.0,<2.0.0",
    "tenacity>=9.0.0,<10.0.0",
    "threadpoolctl>=3.5.0,<4.0.0",
    "tzdata>=2024.2",
    "urllib3>=2.2.3,<3.0.0",
    "zipp>=3.21.0,<4.0.0",
    "plantuml-markdown>=3.5.0,<4.0.0",
    "diskcache>=5.6.3,<6.0.0",
]

# ============================================================================
# OPTIONAL DEPENDENCIES (INSTALLATION GROUPS)
# ============================================================================
[project.optional-dependencies]

# Development tools
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "pytest-benchmark>=4.0.0",
    "black>=23.0.0,<25.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<8.0.0",
    "mypy>=1.0.0,<2.0.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0,<2.0.0",
    "mkdocs-material>=9.5.0,<10.0.0",
    "mkdocstrings[python]>=0.24.0,<1.0.0",
    "pymdown-extensions>=10.0,<11.0",
    "mkdocs-minify-plugin>=0.8.0,<1.0.0",
    "plantuml-markdown>=3.5.0,<4.0.0",
    "mike>=2.0.0,<3.0.0",
]

# Production server
server = [
    "gunicorn>=21.2.0,<22.0.0",
    "gevent>=24.2.1,<25.0.0",
    "greenlet>=3.0.3,<4.0.0",
    "waitress>=2.1.0,<3.0.0",
]

# Cache system (Phase 1)
cache = [
    "redis>=5.2.0,<6.0.0",
    "celery>=5.4.0,<6.0.0",
    "asyncio-pool>=0.6.0,<1.0.0",
    "aiofiles>=23.2.1,<24.0.0",
    "lz4>=4.3.3,<5.0.0",
    "blosc2>=2.7.1,<3.0.0",
    "psutil>=6.1.0,<7.0.0",
    "memory-profiler>=0.61.0,<1.0.0",
    "dill>=0.3.8,<1.0.0",
    "cloudpickle>=3.1.0,<4.0.0",
]

# Profiling tools
profiling = [
    "py-spy>=0.3.14,<1.0.0",
    "memory-profiler>=0.61.0,<1.0.0",
    "psutil>=6.1.0,<7.0.0",
]

# Advanced compression (optional)
compression = [
    "python-snappy>=0.7.3,<1.0.0",
    "zstandard>=0.23.0,<1.0.0",
]

# All extras for complete development environment
all = [
    "biorempp-web[dev,docs,server,cache,profiling,compression]",
]

# ============================================================================
# PACKAGE CONFIGURATION
# ============================================================================
[tool.setuptools]
packages = [
    "src",
    "src.domain",
    "src.application",
    "src.presentation",
    "src.infrastructure",
    "src.shared",
    "src.data_tables",
    "src.assets",
    "config",
]

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json", "*.css", "*.js"]

# ============================================================================
# TOOL CONFIGURATIONS
# ============================================================================

# Black formatter
[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | site
  | deprecated_docs_biorempp_web
)/
'''

# isort import sorter
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/venv/*", "*/.venv/*", "*/site/*"]

# Flake8 (moved from .flake8)
[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "E266", "E501", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "build",
    "dist",
    "*.egg-info",
    "site",
]
max-complexity = 10

# mypy type checker
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

# pytest configuration (moved from pytest.ini)
[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "domain: Domain layer tests",
]

# Coverage configuration (moved from pytest.ini)
[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
    "*/venv/*",
    "*/.venv/*",
    "*/deprecated_docs_biorempp_web/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
